+++ src/setup.js
@@ -1,6 +1,9 @@
 require('../configure')();
 
-const web3 = require('./configure-web3').getWeb3();
+const web3 = require('./config/web3').getWeb3();
+require('@openzeppelin/contract-loader/lib/configure').set(web3);
+
+const { load } = require('@openzeppelin/contract-loader');
 
 const chai = require('chai');
 const BN = web3.utils.BN;
@@ -10,4 +13,5 @@ chai.use(require('chai-bn')(BN));
 module.exports = {
   web3,
   BN,
+  load,
 };