+++ src/singletons.js
@@ -1,28 +1,20 @@
 const { web3 } = require('./setup');
 const ether = require('./ether');
 const send = require('./send');
+const tryRequire = require('try-require');
 
 const {
   ERC1820_REGISTRY_ABI,
   ERC1820_REGISTRY_ADDRESS,
-  ERC1820_REGISTRY_BYTECODE,
   ERC1820_REGISTRY_DEPLOY_TX,
 } = require('./data');
-
-const contract = require('truffle-contract');
-
-const ERC1820RegistryArtifact = contract({
-  abi: ERC1820_REGISTRY_ABI,
-  unlinked_binary: ERC1820_REGISTRY_BYTECODE, /* eslint-disable-line camelcase */
-});
-ERC1820RegistryArtifact.setProvider(web3.currentProvider);
 
 async function ERC1820Registry (funder) {
   // Read https://eips.ethereum.org/EIPS/eip-1820 for more information as to how the ERC1820 registry is deployed to
   // ensure its address is the same on all chains.
 
   if ((await web3.eth.getCode(ERC1820_REGISTRY_ADDRESS)).length > '0x0'.length) {
-    return ERC1820RegistryArtifact.at(ERC1820_REGISTRY_ADDRESS);
+    return getDeployedERC1820Registry();
   }
 
   // 0.08 ether is needed to deploy the registry, and those funds need to be transferred to the account that will deploy
@@ -31,7 +23,39 @@ async function ERC1820Registry (funder) {
 
   await web3.eth.sendSignedTransaction(ERC1820_REGISTRY_DEPLOY_TX);
 
-  return ERC1820RegistryArtifact.at(ERC1820_REGISTRY_ADDRESS);
+  return getDeployedERC1820Registry();
+}
+
+async function getDeployedERC1820Registry () {
+  const environment = require('./config/environment').getEnviroment();
+
+  if (environment === 'truffle') {
+    let truffleContract;
+    [ '@truffle/contract', 'truffle-contract' ].forEach(pkg => {
+      if (truffleContract === undefined) {
+        truffleContract = tryRequire(pkg);
+      }
+    });
+
+    if (truffleContract === undefined) {
+      throw new Error(`\
+Current environment is 'truffle', but found no truffle contract abstraction package.
+Install it via:
+  npm install @truffle/contract
+`);
+    }
+
+    const contractAbstraction = truffleContract({ abi: ERC1820_REGISTRY_ABI });
+    contractAbstraction.setProvider(web3.currentProvider);
+
+    return contractAbstraction.at(ERC1820_REGISTRY_ADDRESS);
+
+  } else if (environment === 'web3') {
+    return new web3.eth.Contract(ERC1820_REGISTRY_ABI, ERC1820_REGISTRY_ADDRESS);
+
+  } else {
+    throw new Error(`Unknown environment: '${environment}'`);
+  }
 }
 
 module.exports = {