+++ test/src/expectEvent.truffle.test.js
@@ -6,7 +6,7 @@ const expectEvent = require('../../src/expectEvent');
 const EventEmitter = artifacts.require('EventEmitter');
 const IndirectEventEmitter = artifacts.require('IndirectEventEmitter');
 
-describe('expectEvent', function () {
+contract('expectEvent (truffle contracts)', function ([deployer]) {
   beforeEach(async function () {
     this.constructionValues = {
       uint: 42,
@@ -19,6 +19,8 @@ describe('expectEvent', function () {
       this.constructionValues.boolean,
       this.constructionValues.string
     );
+
+    this.secondEmitter = await IndirectEventEmitter.new();
   });
 
   describe('inConstructor', function () {
@@ -67,18 +69,18 @@ describe('expectEvent', function () {
     });
   });
 
-  describe('inLogs', function () {
+  describe('default', function () {
     describe('with no arguments', function () {
       beforeEach(async function () {
-        ({ logs: this.logs } = await this.emitter.emitArgumentless());
+        this.receipt = await this.emitter.emitArgumentless();
       });
 
       it('accepts emitted events', function () {
-        expectEvent.inLogs(this.logs, 'Argumentless');
+        expectEvent(this.receipt, 'Argumentless');
       });
 
       it('throws if an unemitted event is requested', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent')).to.throw();
+        expect(() => expectEvent(this.receipt, 'UnemittedEvent')).to.throw();
       });
     });
 
@@ -86,108 +88,108 @@ describe('expectEvent', function () {
       context('short uint value', function () {
         beforeEach(async function () {
           this.value = 42;
-          ({ logs: this.logs } = await this.emitter.emitShortUint(this.value));
+          this.receipt = await this.emitter.emitShortUint(this.value);
         });
 
         it('accepts emitted events with correct BN', function () {
-          expectEvent.inLogs(this.logs, 'ShortUint', { value: new BN(this.value) });
+          expectEvent(this.receipt, 'ShortUint', { value: new BN(this.value) });
         });
 
         it('throws if an emitted event with correct JavaScript number is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'ShortUint', { value: this.value })).to.throw();
+          expect(() => expectEvent(this.receipt, 'ShortUint', { value: this.value })).to.throw();
         });
 
         it('throws if an emitted event with correct BN and incorrect name is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'ShortUint', { number: new BN(this.value) })).to.throw();
+          expect(() => expectEvent(this.receipt, 'ShortUint', { number: new BN(this.value) })).to.throw();
         });
 
         it('throws if an unemitted event is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.value })).to.throw();
+          expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.value })).to.throw();
         });
 
         it('throws if an incorrect value is passed', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'ShortUint', { value: 23 })).to.throw();
+          expect(() => expectEvent(this.receipt, 'ShortUint', { value: 23 })).to.throw();
         });
       });
 
       context('short int value', function () {
         beforeEach(async function () {
           this.value = -42;
-          ({ logs: this.logs } = await this.emitter.emitShortInt(this.value));
+          this.receipt = await this.emitter.emitShortInt(this.value);
         });
 
         it('accepts emitted events with correct BN', function () {
-          expectEvent.inLogs(this.logs, 'ShortInt', { value: new BN(this.value) });
+          expectEvent(this.receipt, 'ShortInt', { value: new BN(this.value) });
         });
 
         it('throws if an emitted event with correct JavaScript number is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'ShortInt', { value: this.value })).to.throw();
+          expect(() => expectEvent(this.receipt, 'ShortInt', { value: this.value })).to.throw();
         });
 
         it('throws if an unemitted event is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.value })).to.throw();
+          expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.value })).to.throw();
         });
 
         it('throws if an incorrect value is passed', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'ShortInt', { value: -23 })).to.throw();
+          expect(() => expectEvent(this.receipt, 'ShortInt', { value: -23 })).to.throw();
         });
       });
 
       context('long uint value', function () {
         beforeEach(async function () {
           this.bigNumValue = new BN('123456789012345678901234567890');
-          ({ logs: this.logs } = await this.emitter.emitLongUint(this.bigNumValue));
+          this.receipt = await this.emitter.emitLongUint(this.bigNumValue);
         });
 
         it('accepts emitted events with correct BN', function () {
-          expectEvent.inLogs(this.logs, 'LongUint', { value: this.bigNumValue });
+          expectEvent(this.receipt, 'LongUint', { value: this.bigNumValue });
         });
 
         it('throws if an unemitted event is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.bigNumValue })).to.throw();
+          expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.bigNumValue })).to.throw();
         });
 
         it('throws if an incorrect value is passed', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'LongUint', { value: 2300 })).to.throw();
+          expect(() => expectEvent(this.receipt, 'LongUint', { value: 2300 })).to.throw();
         });
       });
 
       context('long int value', function () {
         beforeEach(async function () {
           this.bigNumValue = new BN('-123456789012345678901234567890');
-          ({ logs: this.logs } = await this.emitter.emitLongInt(this.bigNumValue));
+          this.receipt = await this.emitter.emitLongInt(this.bigNumValue);
         });
 
         it('accepts emitted events with correct BN', function () {
-          expectEvent.inLogs(this.logs, 'LongInt', { value: this.bigNumValue });
+          expectEvent(this.receipt, 'LongInt', { value: this.bigNumValue });
         });
 
         it('throws if an unemitted event is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.bigNumValue })).to.throw();
+          expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.bigNumValue })).to.throw();
         });
 
         it('throws if an incorrect value is passed', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'LongInt', { value: -2300 })).to.throw();
+          expect(() => expectEvent(this.receipt, 'LongInt', { value: -2300 })).to.throw();
         });
       });
 
       context('address value', function () {
         beforeEach(async function () {
           this.value = '0x811412068E9Fbf25dc300a29E5E316f7122b282c';
-          ({ logs: this.logs } = await this.emitter.emitAddress(this.value));
+          this.receipt = await this.emitter.emitAddress(this.value);
         });
 
         it('accepts emitted events with correct address', function () {
-          expectEvent.inLogs(this.logs, 'Address', { value: this.value });
+          expectEvent(this.receipt, 'Address', { value: this.value });
         });
 
         it('throws if an unemitted event is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.value })).to.throw();
+          expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.value })).to.throw();
         });
 
         it('throws if an incorrect value is passed', function () {
           expect(() =>
-            expectEvent.inLogs(this.logs, 'Address', { value: '0x21d04e022e0b52b5d5bcf90b7f1aabf406be002d' })
+            expectEvent(this.receipt, 'Address', { value: '0x21d04e022e0b52b5d5bcf90b7f1aabf406be002d' })
           ).to.throw();
         });
       });
@@ -195,38 +197,38 @@ describe('expectEvent', function () {
       context('boolean value', function () {
         beforeEach(async function () {
           this.value = true;
-          ({ logs: this.logs } = await this.emitter.emitBoolean(this.value));
+          this.receipt = await this.emitter.emitBoolean(this.value);
         });
 
         it('accepts emitted events with correct address', function () {
-          expectEvent.inLogs(this.logs, 'Boolean', { value: this.value });
+          expectEvent(this.receipt, 'Boolean', { value: this.value });
         });
 
         it('throws if an unemitted event is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.value })).to.throw();
+          expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.value })).to.throw();
         });
 
         it('throws if an incorrect value is passed', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'Boolean', { value: false })).to.throw();
+          expect(() => expectEvent(this.receipt, 'Boolean', { value: false })).to.throw();
         });
       });
 
       context('string value', function () {
         beforeEach(async function () {
           this.value = 'OpenZeppelin';
-          ({ logs: this.logs } = await this.emitter.emitString(this.value));
+          this.receipt = await this.emitter.emitString(this.value);
         });
 
         it('accepts emitted events with correct string', function () {
-          expectEvent.inLogs(this.logs, 'String', { value: this.value });
+          expectEvent(this.receipt, 'String', { value: this.value });
         });
 
         it('throws if an unemitted event is requested', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.value })).to.throw();
+          expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.value })).to.throw();
         });
 
         it('throws if an incorrect value is passed', function () {
-          expect(() => expectEvent.inLogs(this.logs, 'String', { value: 'ClosedZeppelin' })).to.throw();
+          expect(() => expectEvent(this.receipt, 'String', { value: 'ClosedZeppelin' })).to.throw();
         });
       });
 
@@ -234,38 +236,38 @@ describe('expectEvent', function () {
         context('with non-null value', function () {
           beforeEach(async function () {
             this.value = '0x12345678';
-            ({ logs: this.logs } = await this.emitter.emitBytes(this.value));
+            this.receipt = await this.emitter.emitBytes(this.value);
           });
 
           it('accepts emitted events with correct bytes', function () {
-            expectEvent.inLogs(this.logs, 'Bytes', { value: this.value });
+            expectEvent(this.receipt, 'Bytes', { value: this.value });
           });
 
           it('throws if an unemitted event is requested', function () {
-            expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.value })).to.throw();
+            expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.value })).to.throw();
           });
 
           it('throws if an incorrect value is passed', function () {
-            expect(() => expectEvent.inLogs(this.logs, 'Bytes', { value: '0x123456' })).to.throw();
+            expect(() => expectEvent(this.receipt, 'Bytes', { value: '0x123456' })).to.throw();
           });
         });
 
         context('with null value', function () {
           beforeEach(async function () {
             this.value = '0x';
-            ({ logs: this.logs } = await this.emitter.emitBytes(this.value));
+            this.receipt = await this.emitter.emitBytes(this.value);
           });
 
           it('accepts emitted events with correct bytes', function () {
-            expectEvent.inLogs(this.logs, 'Bytes', { value: null });
+            expectEvent(this.receipt, 'Bytes', { value: null });
           });
 
           it('throws if an unemitted event is requested', function () {
-            expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: null })).to.throw();
+            expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: null })).to.throw();
           });
 
           it('throws if an incorrect value is passed', function () {
-            expect(() => expectEvent.inLogs(this.logs, 'Bytes', { value: '0x123456' })).to.throw();
+            expect(() => expectEvent(this.receipt, 'Bytes', { value: '0x123456' })).to.throw();
           });
         });
       });
@@ -276,32 +278,33 @@ describe('expectEvent', function () {
         this.uintValue = new BN('123456789012345678901234567890');
         this.booleanValue = true;
         this.stringValue = 'OpenZeppelin';
-        ({ logs: this.logs } =
-          await this.emitter.emitLongUintBooleanString(this.uintValue, this.booleanValue, this.stringValue));
+        this.receipt = await this.emitter.emitLongUintBooleanString(
+          this.uintValue, this.booleanValue, this.stringValue
+        );
       });
 
       it('accepts correct values', function () {
-        expectEvent.inLogs(this.logs, 'LongUintBooleanString', {
+        expectEvent(this.receipt, 'LongUintBooleanString', {
           uintValue: this.uintValue, booleanValue: this.booleanValue, stringValue: this.stringValue,
         });
       });
 
       it('throws with correct values assigned to wrong arguments', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'LongUintBooleanString', {
+        expect(() => expectEvent(this.receipt, 'LongUintBooleanString', {
           uintValue: this.booleanValue, booleanValue: this.uintValue, stringValue: this.stringValue,
         })).to.throw();
       });
 
       it('throws when any of the values is incorrect', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'LongUintBooleanString', {
+        expect(() => expectEvent(this.receipt, 'LongUintBooleanString', {
           uintValue: 23, booleanValue: this.booleanValue, stringValue: this.stringValue,
         })).to.throw();
 
-        expect(() => expectEvent.inLogs(this.logs, 'LongUintBooleanString', {
+        expect(() => expectEvent(this.receipt, 'LongUintBooleanString', {
           uintValue: this.uintValue, booleanValue: false, stringValue: this.stringValue,
         })).to.throw();
 
-        expect(() => expectEvent.inLogs(this.logs, 'LongUintBooleanString', {
+        expect(() => expectEvent(this.receipt, 'LongUintBooleanString', {
           uintValue: this.uintValue, booleanValue: this.booleanValue, stringValue: 'ClosedZeppelin',
         })).to.throw();
       });
@@ -311,21 +314,21 @@ describe('expectEvent', function () {
       beforeEach(async function () {
         this.uintValue = 42;
         this.booleanValue = true;
-        ({ logs: this.logs } = await this.emitter.emitLongUintAndBoolean(this.uintValue, this.booleanValue));
+        this.receipt = await this.emitter.emitLongUintAndBoolean(this.uintValue, this.booleanValue);
       });
 
       it('accepts all emitted events with correct values', function () {
-        expectEvent.inLogs(this.logs, 'LongUint', { value: new BN(this.uintValue) });
-        expectEvent.inLogs(this.logs, 'Boolean', { value: this.booleanValue });
+        expectEvent(this.receipt, 'LongUint', { value: new BN(this.uintValue) });
+        expectEvent(this.receipt, 'Boolean', { value: this.booleanValue });
       });
 
       it('throws if an unemitted event is requested', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.uintValue })).to.throw();
+        expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.uintValue })).to.throw();
       });
 
       it('throws if incorrect values are passed', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'LongUint', { value: 23 })).to.throw();
-        expect(() => expectEvent.inLogs(this.logs, 'Boolean', { value: false })).to.throw();
+        expect(() => expectEvent(this.receipt, 'LongUint', { value: 23 })).to.throw();
+        expect(() => expectEvent(this.receipt, 'Boolean', { value: false })).to.throw();
       });
     });
 
@@ -333,38 +336,36 @@ describe('expectEvent', function () {
       beforeEach(async function () {
         this.firstUintValue = 42;
         this.secondUintValue = 24;
-        ({ logs: this.logs } = await this.emitter.emitTwoLongUint(this.firstUintValue, this.secondUintValue));
+        this.receipt = await this.emitter.emitTwoLongUint(this.firstUintValue, this.secondUintValue);
       });
 
       it('accepts all emitted events of the same type', function () {
-        expectEvent.inLogs(this.logs, 'LongUint', { value: new BN(this.firstUintValue) });
-        expectEvent.inLogs(this.logs, 'LongUint', { value: new BN(this.secondUintValue) });
+        expectEvent(this.receipt, 'LongUint', { value: new BN(this.firstUintValue) });
+        expectEvent(this.receipt, 'LongUint', { value: new BN(this.secondUintValue) });
       });
 
       it('throws if an unemitted event is requested', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'UnemittedEvent', { value: this.uintValue })).to.throw();
+        expect(() => expectEvent(this.receipt, 'UnemittedEvent', { value: this.uintValue })).to.throw();
       });
 
       it('throws if incorrect values are passed', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'LongUint', { value: new BN(41) })).to.throw();
-        expect(() => expectEvent.inLogs(this.logs, 'LongUint', { value: 24 })).to.throw();
+        expect(() => expectEvent(this.receipt, 'LongUint', { value: new BN(41) })).to.throw();
+        expect(() => expectEvent(this.receipt, 'LongUint', { value: 24 })).to.throw();
       });
     });
 
     describe('with events emitted by an indirectly called contract', function () {
       beforeEach(async function () {
-        this.secondEmitter = await IndirectEventEmitter.new();
-
         this.value = 'OpenZeppelin';
-        ({ logs: this.logs } = await this.emitter.emitStringAndEmitIndirectly(this.value, this.secondEmitter.address));
+        this.receipt = await this.emitter.emitStringAndEmitIndirectly(this.value, this.secondEmitter.address);
       });
 
       it('accepts events emitted by the directly called contract', function () {
-        expectEvent.inLogs(this.logs, 'String', { value: this.value });
+        expectEvent(this.receipt, 'String', { value: this.value });
       });
 
       it('throws when passing events emitted by the indirectly called contract', function () {
-        expect(() => expectEvent.inLogs(this.logs, 'IndirectString', { value: this.value })).to.throw();
+        expect(() => expectEvent(this.receipt, 'IndirectString', { value: this.value })).to.throw();
       });
     });
   });
@@ -373,8 +374,6 @@ describe('expectEvent', function () {
     describe('when emitting from called contract and indirect calls', function () {
       context('string value', function () {
         beforeEach(async function () {
-          this.secondEmitter = await IndirectEventEmitter.new();
-
           this.value = 'OpenZeppelin';
           const { receipt } = await this.emitter.emitStringAndEmitIndirectly(this.value, this.secondEmitter.address);
           this.txHash = receipt.transactionHash;
@@ -441,6 +440,21 @@ describe('expectEvent', function () {
             ));
           });
         });
+      });
+    });
+
+    describe('with non-unique event names', function () {
+      it('throws', async function () {
+        const { transactionHash } = await this.emitter.emitRepeated('0x');
+        await assertFailure(expectEvent.inTransaction(transactionHash, EventEmitter, 'Repeated'));
+      });
+    });
+
+    describe('with non-existing event names', function () {
+      it('throws', async function () {
+        // Which function we call is not important for this test
+        const { transactionHash } = await this.emitter.emitArgumentless();
+        await assertFailure(expectEvent.inTransaction(transactionHash, EventEmitter, 'Nonexistant'));
       });
     });
   });